In WebGLRenderer transparent materials are now handled in the same pass as blended materials (resulting in better fake transparency ;)
The order of render passes is now following:

1. opaque materials with normal blending

2. opaque materials with additive blending
3. opaque materials with subtractive blending

4. transparent materials with additive blending
5. transparent materials with subtractive blending
6. transparent materials with normal blending

With growing number of passes it's possible it may be worth to do some pre-sorting, though it would have to be tested.

Current way is fairly cheap as if there are no corresponding materials in the pass, for loops just blast through (potentially many very cheap operations).

Pre-sorting would mean creating arrays on the fly in each frame (one expensive operation with effect on garbage collection, which especially Firefox is currently very bad at).