Add simple node.js dependency-less HTTP Server to utils
Its sometimes easier to run node.js than python on some computers (eg. on a fresh installation of windows). Although previously discussed in #534, I thought it'd be nice to have a simple web server in jss since a python-less build system is already in place.

This could be an alternative to `python -m SimpleHTTPServer`

To use, cd to the base directory and run
`node utils/simplehttpserver.js <port number>`