Fixed editor drag/drop import files in Chrome and Firefox ( Fixes issue #3028 ).
The DOM event `dragover` needs to be handled as well as `drag`.

The Worker callback's invocation of `parseFile` provided incorrect arguments. To address this, I've extracted the logic that adds a data object to the scene into its own function: `addFileObject`.

This works as expected on Google Chrome 24.0.1312.57 and Firefox 18.0.1 on Linux (Ubuntu 12.04).