Improved home page
List of changes:
The links on the side have the same font family, but the main content
has a new font family.

I removed the br tags in the navigation and use css instead.
There is a background behind each link the the navigation.
The background gets darker when you hover your mouse over it ( only in
webkit.  I haven't added the css for the other web browsers).

I have the navigation links arranged in groups, which there is extra
space between them.

I added a image for the download link.

I created an about section on the top of the page.
There is three getting started links that make it easier to get started.
There is a footer on the bottom of the page.
h2 tags have larger font size and a new color.
There is more space around each featured project and it shows the
title.  When you hover your mouse over the featured project the
background darkens and the color transitions to white.