On the fly sorting of object hierarchy in the UI
Without this fix, the object hierarchy will not be updated in the UI unless you reload the editor.

Worth noting that due to the way UI.FancySelect.prototype.setOptions works, we have to swap around the IDs of ThreeJS objects that are meant to be listed. (ID, not UUID fortunately!)
There is no adverse effect to doing this in the editor app since it never relies on a given object having a given id.