Dolly translates camera rather than scales its distance
At least on Macs, trackpad swipes already trigger a sequence of
mousewheel events with smoothly damped deltas, continuing after the
fingers have been lifted from the pad. Unfortunately, the intuitive,
smoothly damped dollying motion that this would cause is canceled by the
exponential zooming behavior (multiplying or dividing the current
distance by a fixed value).

It feels much better to dolly in and out in a linear manner, as this
patch does. The change is especially noticable if
(maxDistance-minDistance) is finite.

Note that different browsers produce mousewheel deltas with different
scales. This patch copies some code from TrackballControls to normalize
to a common scale.