Add penumbra angle to spotLight
Add an additional option to `THREE.SpotLight` called `penumbraAngle`.
This angle will be added to the angle of the spotlight to create a
"falloff" cone. The area within both cones is lit as normal. The area
outside of the falloff code is not lit. The area between the inner and
falloff cones is smoothly faded away as it approaches the edge of the
falloff cone.