AutoReverse flag for THREE.ShapeGeometry options
When I create a `THREE.ShapeGeometry` from a list of points in clockwise order because I want the result to have the frontside of my resulting faces in the opposite direction the method automatically reverses the order of my points. It would be useful to have an autoReverse flag in the ShapeGeometry configuration/options to be able to control this behavior.