Morph animation name scheme
Currently the exported morph targets have a hardcoded "animation_" prefix, this change names the animations according to timeline markers in Blender. The markers name all animation frames from the frame of the marker up to the next marker. If there are no markers, the animation frames are named as they were.
Making a character rig I missed this functionality, which gives me the ability to name animations (idle, walk, run, etc.)