Fixed `fov` parsing
Changed `cparams.yfov` to `parseFloat(cparams.xfov || cparams.yfov)`.

Collada files exported from Cinema 4D use an `<xfov>` tag to specify the FOV of cameras within a scene. Both `xfov` and `yfov`are parsed in the `parseOptics` function, but they are not correctly passed onto the `PerspectiveCamera` constructor, so new instances default to an FOV of `50`.

This update resolves this issue.