Safe EventDispatcher
This is just a suggestion.

The current implementation of EventDispatcher is intrusive. It sticks `_listeners` on
object it's being added to. If that object already has a `_listeners` your out of luck.

Above is a non intrusive implementation for your consideration. The `listeners`
collection is not public whatsoever.

I [tested them on jsperf and it's the same speed on chrome and significantly faster
on Firefox](http://jsperf.com/event-dispatching)

I didn't expose `dispatchEvent` to the public API of `TypedGeometry` because it
seemed to me users of `TypedGeometry` shouldn't be able to dispatch events but
of course that could be exposed too.

Anyway, just passing it on as a suggestion. I'm learning JS as I go and finding that what I learned
at Google was wrong :P  They use their Closure Compiler and they basically
treat JavaScript like Java and expect the compiler to make things semi safe but
JavaScript can be safe all by itself so I'm slowly learing more JavaScripty
ways.