Simplify edge removal heuristic
Minor weak for non-manifold scenario.

Anytime an edge is non-manifold (more than two faces sharing an edge, e.g. a 'T'), always keep the edge in the EdgesGeometry.

With the previous commit, when a third co-planar face was found, the edge was marked co-planar. I'm not sure which is right but this algorithm is a bit simpler.