add updateVertices method
This is a very small, simple method that would make it a bit simpler to update the vertices on a geometry object. It takes an iterator as an argument, and the passed iterator should manipulate the the values for each vertex. 

For example, if you wanted to add just a bit of randomness to an objects structure, you could do:

```js
geometry.updateVertices(function(vertex){
    vertex.x += (Math.random() * 2 - 1);
    vertex.y += (Math.random() * 2 - 1);
    vertex.z += (Math.random() * 2 - 1);
});
```