Getting rid of all the computed dependencies
This commit removes all the computed properties on the THREE object, using lookup objects.

- this allows static dependency analysis to work
- this allows tools to perform source conversion to any module format (commonJS, AMD, es6, etc.)

In the loaders case, this also adds an extra layer of protection agains setting a wrong type on the json object.
Previously, setting, for instance, "Object3D" as the material type on a json file you load with MaterialLoader would have instantiated a new Object3D.