Refactor WebGLRenderer fog
There has been a lot of talk on breaking down the renderer in smaller
components, but I have not seen a lot of detail on how to accomplish
this refactoring. Would this be a pattern you would consider?

I've only touched three small part of the renderer - but in case the
changelog is too confusing, here is an overview of the approach.