Added Support in Font.Utils  for Type1 fonts
According to font specifications, both Type1 and TrueType require that
the outer boundary of a glyph be drawn in a certain direction (they
happen to be opposite from each other).  The TrueType font file
specification tells us that contours that need to be filled black must
have a clockwise direction. If we want to make a white area inside an
existing contour we must make the direction of the new contour counter
clockwise, while it is opposite case in Type1.

Added a check and flag (isCCw) for the same. while creating shapes.