use filepath() instead of filepath_from_user()
When using the blender exporter, enabling the texture setting would cause the exporter to crash.

The crash was due to calling a method which doesn't exist: filepath_from_user(). A grep of the repo showed the method is not defined anywhere and replacing it with the filepath() yields the desired result.