With WestLangley suggestions
Add a new variable to avoid instantiation of a Vector3 in each call to
THREE.Triangle.barycoordFromPoint().

Defer calculus after the near plane / far plane test. Need to do a copy
of intersectionPoint due to the applyMatrix4.

Test with a PlaneBufferGeometry and a BoxGeometry. Didn’t test with
indexed BufferGeometry (don’t have any in my program) but I’m quite
comfortable the uvA, uvB and uvC are the right one in that case.