Drag gizmo handles with left mouse button only
In the Three.js editor, the right mouse button is used to pan the view and the middle mouse button is used to zoom in / out so it doesn't make sense that they'd also grab the transform handles.

You can reproduce this bug by holding middle click on an arrow of the translation gizmo and moving up/down. The view will zoom in/out while the dragged object moves too, which is surprising and not useful.

This PR limits dragging to the left mouse button, fixing the issue.