Remove redundant loop from dispatchEvent
The 'dispatchEvent' method of 'THREE.EventDispatcher' ran two loops
to execute callbacks. The first copied references to callback
functions from the listener array to another array, and then the
second loop would call the callbacks from this second array. This
is completely unnecessary, therefore the second loop was removed
and the callbacks are now accessed from the listener array when
executed.

A random sample of 12 examples seem functional after the patch.