Updates all light documentation to use color, not hex
Addresses #7707. It's confusing to have a Color data type but then tell
users they can only use a primitive value that *represets* a Color, even
though it's converted to a Color under the hood.

Note that the Material API still encourages the use of hex, which should
also be updated for consistency.