Fix for incorrect diffuse color being used.
I found this issue testing the blender exporter. Basically randomly the
wrong diffuse color would be used. This was because the order of the
json properties would be different, and the DbgColor would overwrite the
colorDiffuse, this patch checks if the color has already been set and
doesn't set it to DbgColor.