Custom UV on LatheGeometry without custom geometry
Example of custom UV in the Editor before introduction of custom
geometries on ObjectLoader.

LatheGeometry.j and ObjectLoader.js are modified to give an "easy"
access to the data serialised.
The same kind of modification would be required for all geometry.
The switch case of parseGeometries maybe greetly simplified because only
CubeGeometry, BufferGeometry and Geometry would need a specific code.