Removed duplicated test, and rearranged replaceAt
I had a newChild===oldChild test in replace as an optimisation, because I was thinking that replacing an object with itself is normal operation. But I think it is not a normal for replace as for replaceAt, and therefore I removed the test from replace.

I rearranged the statements in the end of replaceAt, so that the actual changes are at the bottom, and the events are triggered right after that.