Set renderer to correct resolution when entering VR mode
This ensures that the scene always renders with the best resolution for
the HMD in use. (Will get a 1:1 pixel ratio) For performance reasons
developers may want to render at a lower than recommended resolution. To
do so set VREffect.bufferScale to a lower values (like 0.75).

Will restore the renderer to the previously set resolution when the user
exits VR mode.