Last spline coordinates not stored in actions
In the actions array of the Path, the last elements of the args array
need to be the x and y of the last point of the segment added to the
path.
Not the case with splineThru => cause an exception if you add an other
segment after the pline (without a moveTo).
Used the same trick as absellipse does.