Add ray in ViveControllerObject
I suggest to add a ray in the ViveControllerObject as simple interaction concept. This ray is an easy understandable gesture (as shown by Valve's Steam Overlay) and mandatory for many interactions. It can be used for ray intersection, but still belongs to the visualization.

As an alternative, I suggest to extend the ViveController example. If requested, I will create a pull request for this as well.