Update dat.gui
dat.gui had an issue where if you call .listen() on dropdown
controls they don't work in Chrome on OSX

You can see an example here
http://threejs.org/examples/webgl_postprocessing_ssao.html

On Chrome OSX try to change the renderMode

There are 2 fixes. One is to remove the call to `listen`.
The problem with this solution is that since things
work on say, Windows someone using dat.gui has no idea
their stuff doesn't work else where.

The other is to get a version of dat.gui that doesn't
have the issue as it was [recently fixed](https://github.com/dataarts/dat.gui/issues/101)