Partial revert of some WebVR 1.1 changes
They didn't produce correct eye translations in all cases (specifically
standing scenes.) This reverted version may not be 100% accurate on
future hardware (if the eyes have a rotation component, for example) but
it's correct on all current WebVR implementations. That's better than
blatantly wrong in some cases.

Use of the WebVR-provided projection matricies has been left in, since
that was working correctly.