Fix Race Condition inside ImageLoader for base64
Experienced this while using Node. Basically the issue resides when
using `THREE.ObjectLoader` where the images in the JSON are base64.

On my very fast system, it was called the `THREE.LoadingManager`
`itemEnd` before the `itemStart` . Causing the callback `onLoad` for
the `LoadingManager`to never get called.

When using `THREE.ObjectLoader` it tries to call `parseImages` which
has a load callback. When the machine is too quick, it misses the
synchronization when to fire the `onLoad`.

By moving the `scope.manager.itemStart( url )` to the start of the
function, that fixes the synchronization.