Change to allow support for CombinedCamera
The `THREE.CombinedCamera` class can currently not be used with the `THREE.OrbitControls` class because there are `instanceof` checks inside the methods `dollyIn` and `pan`. Instead it would be possible to use the `isPerspectiveCamera` and `isOrthographicCamera` values so the class will be compatible with `THREE.CombinedCamera`. In the `THREE.CombinedCamera` [there are currently similar fields on line](https://github.com/mrdoob/three.js/blob/master/examples/js/cameras/CombinedCamera.js#L50-L51):

`````
this.inPerspectiveMode = true;
this.inOrthographicMode = false;
`````
Those fields could be renamed to `isOrthographicCamera` and `isPerspectiveCamera` and then the combined camera would be compatible with the `THREE.OrbitControls`.