Enable keyboard in OrbitControls
In laptop and without a touchscreen, it's very hard to interact
with OrbitControls.

So I added the arrow keys (when pan is disabled) to allow for rotating
(if the Ctrl key is also pressed the rotation is boosted).

Moreover, the + and - keys of the numeric keypad to zoom.
For laptops without keypad, the A and Z keys works also.