Object bounding box relative to given reference
Added optional arguments to setFromObject and expandByObject that allow generating a bounding box relative to a given reference object's coordinate system. Also added some options regarding invisible objects and materials. I consider this an early sketch of the functionality. It should probably be extended to other methods as well, and the logic for undefined materials is probably not perfect.
Related issue: #8222 (visibility)