Use WebVR 1.1 view matrices correctly in VREffect
This will make VREffect work correctly in cases where the view
transforms are not simply based on the head pose with the eye
translation applied. For example, they may contain eye orientation.
The code is partially based on an earlier attempt to fix this by
Brandon Jones, but is fixed to work in both sitting and standing
examples.

Related to issue #9708