Add support for inverse matrices, per material
It is occasionally required to have access to inverse model and modelView matrices
in a shader, however setting these matrices from client code requires creating a
unique material for every object which needs the matrix. Additionally, computing the
inverse in the shader is expensive and often runs more frequently than necessary (e.g.
per vertex instead of once for the entire mesh).

Other engines (such as Unity) set inverse matrices automatically in the render loop,
which is the goal of this change.

Now, each material may declare that the inverse matrices are required, so a single
material can be used for multiple objects and the uniform inverse matrices will be
set along with the model/modelView matrix.

Also see issue #11162