WebGLProgram: add meshLine to list of parameter names
This is needed to distinguish WebGLProgram instances that use mesh lines
from those that don't. This is done based on a so called program code,
which is, besides other things, based on the list of known parameter
values for the parameters names from the parameter list. If two
materials only differ in whether they use mesh lines and the one that
doesn't is loaded first, then without this change no new program will be
generated if the material with enabled mesh lines is loaded afterwards,
which effectively disables mesh lines for this material.