Apply location to bones and animation data in loader
This should probably be fixed in io_three instead, but this is a working
method for right now.
Note: If io_three does get patched, this patch will break things unless
io_three sets the key 'animations_fixed' at the top level of the json
with any value.
Tested the 3 loader/json examples, along with a couple other scenes
including that scene from the issues post.