Updated raycast function for indexed buffer geometry with multi material
For indexed buffer geometry using multi-materials the current intersect objects functions neglects the material side property ( material.side === undefined ) forcing backface culling. This causes faces to be missed regardless of if the material side property is set to THREE.DoubleSide. The proposed solution alters the functions checkIntersection, checkBufferGeometryIntersection, and Mesh.raycast. The faceIndex of the current triangle being checked is sent to checkBufferGeometryIntersection which forwards it to checkIntersection. In checkIntersection the material is tested to see if it is a multi-material and also if a faceIndex is passed in ( to avoid conflict with non indexed buffer geometry ). If it is a multi material the function then checks the groups associated with the geometry ( if any ) and determines which material is being targeted by the faceIndex. This allows the function to accurately get the material.side value to give a proper backface culling value.