Pass MouseEvents and "old Position" from DragControls / TransformControls back to Event Listener
Am looking to use webgl_geometry_spline_editor as a basis for crafting a CNC pattern editor, and in doing so, am proposing some changes to webgl_geometry_spline_editor, TransformControls.js and DragControls.js in order to permit envisioned functionality associated with manipulating polygons.  In short, this pull request includes the following changes:

1) webgl_geometry_spline_editor now accepts a Ctrl + Left Mouse Click to select / de-select the spline vertices.  Unselected vertices are white, and selected vertices are red.  Without the Ctrl key down, hovering over a vertex is the same as before, attaching the transform controls to the vertex.

2) webgl_geometry_spline_editor now moves all selected vertices as a group.

3) DragControls.js has been modified to pass the MouseEvent object to the user event listeners, in order to permit the use of the MouseEvent by the user application to define more functionality within the application.  In this particular use case, now within webgl_geometry_spline_editor, when holding down the Ctrl +  Left Mouse Click on a spline vertex, the vertex is toggled as a selected object so that multiple vertices can be selected at once, and moved together.

The passing of the MouseEvent object has been applied to all appropriate DragControl dispatchEvents for consistency and potential use by the calling application.

In order to ensure maximum backward compatibility, the MouseEvent object has been added as a property to the dispatchEvent parameter.

4) TransformControls.js has been modified to pass the oldPosition, oldSize, and oldRotationMatrix to the user event listener, so that the relative changes to the object associated with the transform control can be applied to other objects.  Ie, this permits the transform control to be attached to a single object, and now the user application event listener can determine the type and amount of recent change of the transform control, and then apply the same transformational changes to other objects within the scene.

Additionally, like the DragControls, the MouseEvent is also included when the dispatchEvent is called, in order to provide flexibility to the user application to define different actions within the event listener, triggered by special key or mouse events.

In order to ensure maximum backward compatibility, the MouseEvent, oldPosition, oldSize, and oldRotationMatrix objects have been added as a properties to the dispatchEvent parameter.

5) webgl_geometry_spline_editor included a timer to turn off the transform controls, and the comments within the js indicated it was a bit of a mess.  This has been cleaned up and simplified.

R, Jon