TDSLoader: Fix double 'pathing'
See https://github.com/mrdoob/three.js/issues/12730
Fixed issue with extracting and using the url base of the passed url without changing the url itself, see L51.
Also passed down the ```path``` variable down to the ```TextureLoader```. Still don't like the downpassing of the ```path``` but for now it should fix the issue, until we rewrite the complete TDSLoader. 
@stef-levesque
@Mugen87