Solves beginning Orientation issue
Take what you will from this. I tried the fix this before (last week) and something was off (closed it). This seems to be working now.
So if you're in portrait or landscape the same center of the screen in desktop 0,0,0 is presented to the viewer no mater what heading (alphaOffset) they face in the real world, they will start facing the scene front center properly.
Maybe it's just me (my test files). I'm sure you don't like my replacement of "var"s with "let"s and excessive use of the local 'scope' reference vs 'this', so again take what you will from it, but the orientation of the viewer is correct regardless of initial orientation of alpha offset (so far as I can tell).