Resolves Orientation Controls (watered down version)
Here is the watered down version some might prefer.
Original https://github.com/mrdoob/three.js/pull/12959
This puts the lets back to vars (I must note that I disagree with this.)
Removes extra arguments except autoConnect so the developer can still make adjustments to the scoped/exported/non-private values as needed (I'd prefer to just put all hidden functions values on scope).
I considered adding a reset of checksDone to connect method so onDeviceChangeEvent s start to recalculate when they do a re-connect, but didn't.
The developer can call 'disconnect' adjust values (including checksDone if no autoConnect) and call 'connect' again instead of having an autoConnect argument, but this seems wasteful, so to be able to override it via setting autoConnect to false makes sense to me.

All arguments are optional in JS and so it doesn't matter, it's just limiting and causes more hoops to resolve and wasted cycles having to recall them manually after they fire on creation. You have to explicitly set autoConnect to false to proceed with manual settings/fixes and the calling of 'connect', so it's noted that it is still as none breaking as was the extra argument way.
I originally put scope as an argument to hammer home the point that you do NOT need let or var at all really. Since you can use 'call' to set 'this' to your own object then you still don't need it in the args or to call DeviceOrientationControls with 'new' even, so I added that it returns the scope which is what new does for you (also none breaking and clarifing and provides more options as putting scope as an argument).

Then you can call it via call (without adding scope as an argument or using new even.)
ctrls = THREE.DeviceOrientationControls.call( cam, cam );
or
ctrls = THREE.DeviceOrientationControls.call( {}, cam );

[Of course then maybe you don't need 'scope' and can just use 'this', since you know what it is (via call) and the issues with this being sometimes something else aren't as concerning. (This is still an area I'm unsure of [not being a V8 dev] but it is spoken about by Crockford as better to have your scope explicit, if just to avoid potential proxy confusion probably in more complicated class hierarchy situation I think {when I heard this it was not clarified why}. Anyway you guys had 'scope' initially I just used it instead of 'this' after it's declaration as that seemed logical and possibly more efficient being more clearly within the constructors functions scope, and I hate 'this' because it seems like it should be implied as a hoisting target {I guess that original logic is kind of stymied by strict mode thinking?}.)]