fix video-textures for cached/fast video-loads
When creating a video-texture from an existing video-element on the
page, the video might have already reached a playable state. In this
case, the `loadeddata` event has already been fired before the
video-texture was created and the texture will never become renderable.

This can be detected if the video has a `readyState >= 2`
(HAVE_CURRENT_DATA) when initializing the texture.