Prevent navigation if modifiers pressed.
In Windows Ctrl-Click opens a link in a new tab. In MacOS it's Cmd-Click.
But the code here was not respecting the user's wishes. A new tab
would open but the page itself woudl also navigate the same link
which is not what a user is requesting with ctrl-click or cmd-click

Interestingly the demo page did not navigate when pressing Cmd-Click
on MacOS but the URL did change. Now the URL will correctly not
change and the link will open in a new window.