Update DirectionalLight.js
I appreciate the Three.js philosophy of keeping things clean, simple and modular. Yet, I often find myself in situations where I have to create and manipulate a Spherical to modify the direction of DirectionalLights in a sensible way. Unit3D, by comparison (and I like Three.js much better for sure), has angles as the default way of configuring the direction. So I thought I could add it to DirectionalLight as an option. The implementation conserves the radius.