Port UnrealBloomPass to EffectComposer2
The bloom pass is a bit of a special case since the same buffer acts as both an input and output buffer for it, and the pass needs to be able to sample it as a texture. Passing the default framebuffer to the UnrealBloomPass2 doesn't work.

So EffectComposer2 now supports having a pass that's only able to render to a texture. In case it's the final pass when rendering to the default framebuffer, then EffectComposer2 will copy the pass output from the texture to the default framebuffer.