Remove unnecessary unworking lines
The script in its current form isn't working as intended, which can be seen by immobilising the origin of the particles, setting the position randomness to 0, and the turbulence to zero. The lines are removed seem to be the cause of the unintended behaviour, and I see no justification for their presence. Removing them fixes the issue.