Add capability to disable shadow maps per light type
This will save texture uniform slots in case shadow maps are only used for some types of lights but not others.

Right now this only toggles shadow maps in the shader code, same as WebGLShadowMap.enabled. Shadow map related uniforms are kept in the uniform data structures regardless of the settings. To optimize this further in the future, the uniform uploads could be completely disabled as well in case shadow maps are not in use.

A typical use case would be in a scene that has one primary directional light with shadows enabled and a bunch of smaller local point and spot lights without shadows.