Conserve shadow map slots reserved for lights
Only shadow casting lights now get entries in the shadow maps arrays in shaders. Lights are sorted so that shadow casting lights come first, so the index of the light in the shadow maps array still matches the index in the normal light data array.

This conserves texture slots, enabling to use more lights in scenes where shadow maps are enabled. Uniform update performance and shader execution performance should also be a bit better when there's a mix of shadow casting and non-shadowed lights.