FileLoader: Fix error response bodies being cached as results. Fixes #17635.
Until now, the response body was saved to the cache, no matter what the HTTP
status code was.

So a request that returned e.g. HTTP 404 with response body
"No such file" and `onError` being called would, if run a second time,
be found in the cache with "No such file" as the result, and
`onLoad("No such file")` would be called, which is nonsensical.

One common failure was that texture loaders would thus interpret
plain error strings as textures to decode, and failed.