Implement support for texture matrix for uv2
Before this change, we only supported one texture matrix for the
material that was applied to uv data based on the first texture, in
priority order, that needed it.

This change adds equivalent support for uv2 - we now transform second UV
channel with a second UV transform that is grabbed from aoMap or
lightMap (that are the only two textures using it atm).

This fixes glTF rendering for files that use KHR_texture_transform on
the occlusion texture, including ones produced by gltfpack.