Handle getting a detached ArrayBuffer from Cache
ArrayBuffers that are stored in the Cache may not remain valid in case they end up in a detached state. Buffers can enter detached state when transferred to a web worker for example.

This fixes feeding the same URL multiple times to FileLoader when using DRACOLoader, that offloads decoding files to a worker.