Make elements on top that need click events possible
When developing a application with draggable objects that can have a toolbar element on top of it. It is not possible to capture the click event because it is captured by DragControls always. It should only be the case when the click event is comming from the _domElement. In my case I had a CSS2DObject on top of it. The proposed change solves this and is in my opinion always good to check no matter how it is used.