SelectionBox: correctly apply frustum fix
The fix is relevant for both orthographic and perspective camera.

Also, a case where only one coordinate is identical (i.e. the
selection is a line) also causes invalid frustum. This is less
important if you're testing for a single point, because it won't
be inside anyway, but it breaks as soon as you test for
intersection of the box with a bigger object (bounding sphere
or bounding box).