Update docs to indicate how to choose GPU
Add the `powerPreference` parameter to the Getting Started guide so that new users know where to look in case they want to choose which GPU to use in their WebGL context. 

There's lots of misinformation on the web about how to move the rendering onto a discrete GPU or whether that's possible at all (most links point to chrome flags etc). Adding this note in the doc should give users a starting point on where to look. 

I know the "high-performance" flag has some caveats associated with it which is why I included the link to the WebGL spec which clarifies when the user agent might choose to ignore this property. Alternatively, we could add `{ powerPreference: "default" }` just to redundantly indicate that this parameter exists.