WebGLRenderTarget: Add support for swapping textures
This can be used to save GPU memory or to avoid texture copy operations in situations where multiple textures need to be used as render targets but one shared depth/stencil buffer is enough.

The downside is that the WebGL implementation needs to validate the framebuffer again for completeness.

WebGLCubeRenderTarget returns Textures from this function, not CubeTextures.