Update Line3.js
I propose a new method for Line3 to check if two lines are loosely the same, as in one line has the start point of the other line as an end point and vice versa. Currently there is (at least to my knowledge) no way to check wether two lines are loosely the same in a convenient manner (i.e. without for example instantiating a new line with swapped start and end points).