RoughnessMipmapper: Preserve texture UV transform (#21511)
RoughnessMipmapper created a new texture and replaced existing material
textures with the copy - however, the copy didn't inherit the UV
transform properties from the source.

This could be problematic when for example the roughness map came from
the glTF file that used KHR_texture_transform extension. When rendered
with standard three.js material this by itself would not be a problem
since three.js only supports one shared transform for most material maps
but when the user overrode the material with a custom one (which
happens inside model-viewer as far as I can tell), the incorrect
offset/repeat values in the cloned texture could result in incorrect
rendering.