PerspectiveCamera: Add Kooima's Generalized Projection Formulation (#21825)
* Add Kooima's Generalized Projection Matrix Formulation

Addresses Issue #5381

This PR adds an implementation for Kooima's Generalized Projection Matrix Formulation, better known as "the way CAVE rendering works".

In a nutshell, it sets the `projectionMatrix` frustum to exactly frame an arbitrary rectangle.   This is a key operation in rendering portals, CAVEs, and certain kinds of projection based effects.

I find myself porting and reporting this function wherever I go because it's so darn useful.   I'm not certain that this is the right place in the codebase for it, but I'd really love for three.js to have a working implementation for people to use.

Here's a video I made years ago that utilizes the formulation:
https://www.youtube.com/watch?v=90kHhOUzeQc&t=1s

* Switch to const

* Add Portal Example