WebXRManager: Fix Camera Local Space by Reverting #21964
This PR attempts to move XR Camera calculations back into local space by undoing #21964 (which appears to conflate `matrixWorld` and `matrix`).

This enables the XR camera to be parented to another object again without odd behavior.

If this isn't the correct fix; I'd be happy to test alternatives in the context of my application.