A fix for hand tracking in Oculus Browser v14
Oculus Browser on the Quest line of devices supports hand tracking. In version 14, it requires toggling a flag in chrome://flags. The implementation in v14 is defective, however, and the finger joints are rotated in the wrong space.

Position of the joints was correct and the system is generally usable for targeting and manipulation, if not terrifyingly ugly due to "inside out hands" when using the mesh hand model.  This patch corrects the rotation of the joints so that the hand now appears correctly.

Oculus fixed the issue in v15 and also enabled hand tracking by default. Oculus Browser on consumer-licensed devices is currently at v17. However, Oculus Browser for enterprise-licensed devices through Oculus for Business are currently stuck at v14, with no indication from Oculus when they will be upgraded.