Docs: Object3D&SpotLight
About Object3D:
The value type of the children property is Array, but the documents of zh, en, ko all define it as Object3D.

About SpotLight:
In the description of the target attribute, the SpotLight is translated into a DirectionalLight, this should be an unintentional mistake. (the error only occurs in Chinese documents).