WebGLMultipleRenderTargets: Example fix iOS with MSAA
There is another issue only on iOS, some texture types will break in certain cases. I don't have an explanation of the why, I think it's just a bug.

Also, there was an issue with Safari which was fixed in the latest version (15.5) where it was required to clear the canvas before rendering the scene. I only added comments since this issue has been patched, should we add it anyway?