Add normals to BufferGeometry example
I was confused by this for a while because the example doesn't actually render. Eventually I found a more complete example and worked out that the missing piece was normals. I think including normals in the example can help more people like me with a working example to start with.