feat: allow early set foveation
Currently foveation is set to 1.0 on setSession,
and there's no way to set it before,
because it's not saved anywere when session is not active.
So the only way to set it to a non-default value is to call setSession and
call setFoveation afterwards.
That leads to hacks in certain scenarios like https://github.com/pmndrs/react-xr/pull/235.
This PR allows setting it beforehand, while being backwards compatible