Add caching functionality and documentation to TextureLoader class
## Description

This contribution adds caching functionality to the TextureLoader class, along with documenting the added code. The cache stores previously loaded textures and can be used to avoid loading the same texture multiple times, which can greatly improve the performance of the library. The cache size is limited by the maxCacheSize variable and when the cache is full, the least recently used texture will be removed.

## Changes Made
- Added a cache variable of type Map to store previously loaded textures
- Added a maxCacheSize variable to limit the cache size
- Before loading a texture, the code checks if it already exists in the cache
- If it exists, the cached texture is returned instead of loading it again
- If it does not exist, the texture is loaded and added to the cache
- A clearCache() function is added to clear the cache
- Documented the added code to explain the functionality and usage of the cache

## Benefits
- Improves the performance of the library by avoiding loading the same texture multiple times
- Reduces the memory usage by limiting the size of the cache
- Provides a clearCache() function for the user to clear the cache if needed