Linting: Update linter to enforce ES 2018 features (#25470)
* Separate Out eslintConfig

move eslintConfig to its own file to keep package.json clean
modifications and commits not bound to or carry the risk of dependency changes.
This commit is the equivalent configuration as the starting point.

* Update Three.js

Convert ES 2020 imports into 2018.

* Update .eslintrc.json

Set project to ES 2018

* Add Public Interface Export Objects

MathUtils, AnimationUtils, and DataUtils need to be updated to export their namespaces.  Privately, individual functions are exported to support the other parts of the library, publicly they are namespaced with an object wrapper.